# openapi.yaml
...
openapi: 3.1.0
info:
  title: Squarespace Connector API
  version: 1.0.0
  license:
    name: ISC
    identifier: ISC
  description: REST/HTTP surface for the squarespace connector, wrapping the 6 actions declared in connector.yaml (list_products, get_or_adjust_inventory, list_orders, create_order, get_contact, list_contacts) plus a testConnection operational helper. This describes the connector's OWN future HTTP API — not yet deployed, see the servers block — and is distinct from the Squarespace API itself, which this connector calls server-side. API key is sent as a Bearer token in the Authorization header. That Squarespace API key is held server-side and never exposed through this API; the bearer token defined below authenticates callers of THIS API instead. GENERATED by src/generate-openapi.ts from src/schemas/*.json and connector.yaml — do not hand-edit, re-run the generator instead.
servers:
  - url: https://api.example.invalid
    description: Placeholder — this connector has not been deployed yet (.invalid is the RFC 2606 reserved TLD for non-existent placeholder domains). Replace once a real deployment URL exists.
security:
  - bearerAuth: []
paths:
...
  /test-connection:
    post:
      operationId: testConnection
      summary: Verifies that the configured Squarespace API key can reach the API (a live GET /products?limit=1 server-side).
      description: Verifies that the configured Squarespace API key can reach the API (a live GET /products?limit=1 server-side). Not one of connector.yaml's 5 declared actions — an operational helper exposed by src/connector.ts's testConnection() and mcp/server.ts.
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionOutput"
        "400":
          description: Bad request — input failed validation. See ErrorResponse for details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — this API's own bearer token is missing/invalid, or (POST /orders specifically) the underlying Squarespace API key lacks authorization; see connector.yaml's risks entry on create_order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for this connector's own REST/HTTP surface. Separate from the Squarespace API key, which this connector holds server-side (see connector.yaml's auth block) and never exposes to callers of this API.
  schemas:
    ErrorResponse:
      title: ErrorResponse
      description: Normalized error shape. Assembled from connector.yaml's error_handling.normalized_shape, not retyped.
      type: object
      required:
        - code
        - message
        - httpStatus
        - requestId
        - retryable
      properties:
        code:
          type: string
        message:
          type: string
        httpStatus:
          type: integer
        requestId:
          type: string
        retryable:
          type: boolean
    TestConnectionOutput:
      title: TestConnectionResult
      description: Not backed by a src/schemas/*.json file — mirrors src/connector.ts's TestConnectionResult type (INFERRED from that type, not independently schema-verified).
      oneOf:
        - type: object
          required:
            - success
          additionalProperties: false
          properties:
            success:
              const: true
        - type: object
          required:
            - success
            - error
          additionalProperties: false
          properties:
            success:
              const: false
            error:
              description: The raw thrown error value; shape varies by failure mode.
...